refactor: extract the InviteLauncher lane from DispatchOperations
Third launching lane peeled. The invite lane — a human explicitly puts an
agent in a spec's room in full mode (read-only invites are superseded by
engagement) — is a thin lane over the three seams: authorize through
LaunchAdmission, reserve through RunReservation synchronously, then defer the
snapshot payment and launch off the request thread through RunLauncher.

Move startInvite and its helpers (completeInvite, publishInviteFailed,
inviteSnapshot) into InviteLauncher over nine collaborators; the late-bound
message store is injected as a Supplier, and seedRoomDelivery/publish are the
trivial shared helpers each lane carries its own copy of. DispatchOperations
keeps startInvite (both overloads) as delegators, so SailOperations is
unchanged.

Behavior-preserving — InviteLaunchTest passes unchanged, plus three added
cases (no run store, no agent block, empty room) that close the reachable
defensive guards. The two remaining spots — seedRoomDelivery's empty-ledger
guard (a full invite's rendered set is never empty) and its seed catch — are
unreachable without fault injection, gated by a documented per-class rule
(LINE 0.93, METHOD 1.0). mvn clean verify green.
